+++ examples/ios/Sources/BMoEExampleApp.swift
@@ -1,0 +1,8 @@
+import SwiftUI
+
+@main
+struct BMoEExampleApp: App {
+    var body: some Scene {
+        WindowGroup { ContentView() }
+    }
+}